carnot_new/requirements.txt
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.2
autopep8==2.0.2
backcall==0.2.0
black==23.7.0
click==8.1.5
decorator==5.1.1
dill==0.3.6
Django==4.2.3
djangorestframework==3.14.0
executing==1.2.0
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycodestyle==2.10.0
Pygments==2.15.1
pylance==0.5.8
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
wrapt==1.15.0
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.2
autopep8==2.0.2
backcall==0.2.0
black==23.7.0
click==8.1.5
decorator==5.1.1
dill==0.3.6
Django==4.2.3
djangorestframework==3.14.0
executing==1.2.0
gunicorn==21.1.0
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycodestyle==2.10.0
Pygments==2.15.1
pylance==0.5.8
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
whitenoise==6.5.0
wrapt==1.15.0
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.2
autopep8==2.0.2
backcall==0.2.0
black==23.7.0
click==8.1.5
decorator==5.1.1
dill==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
executing==1.2.0
gunicorn==21.1.0
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycodestyle==2.10.0
Pygments==2.15.1
pylance==0.5.8
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
whitenoise==6.5.0
wrapt==1.15.0
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.2
autopep8==2.0.2
backcall==0.2.0
black==23.7.0
click==8.1.5
decorator==5.1.1
dill==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
executing==1.2.0
gunicorn==21.1.0
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycodestyle==2.10.0
Pygments==2.15.1
pylance==0.5.8
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
whitenoise==6.5.0
wrapt==1.15.0
